refactor: extract capture domain into src/capture/ with co-located tests
