+++ src/the-owl.ts
@@ -1,10 +1,12 @@
 import { join } from "node:path";
 import type { Express } from "express";
-import { createCollector } from "./collector";
-import { makeCaptureMiddleware } from "./capture";
+import { createCollector } from "./capture/collector";
+import { createCaptureMiddleware } from "./capture/middleware";
 import { drainToDisk } from "./persist/drain-to-disk";
-import { TEST_NAME_HEADER } from "./headers";
-import { DEFAULT_SANITIZE, normalizeKey, type SanitizeOptions } from "./sanitize";
+import { TEST_NAME_HEADER } from "./capture/constants";
+import { DEFAULT_SANITIZE } from "./capture/constants";
+import { normalizeKey } from "./capture/sanitize";
+import type { SanitizeOptions, ConnectOptions } from "./capture/types";
 
 const collector = createCollector();
 
@@ -12,20 +14,13 @@ export const buildHeaders = (testName: string): Record<string, string> => ({
   [TEST_NAME_HEADER]: testName,
 });
 
-export interface ConnectOptions {
-  /** Override redaction / body-safety defaults (EC2/EC3). */
-  redactHeaders?: Iterable<string>;
-  redactKeys?: Iterable<string>;
-  maxBodyBytes?: number;
-}
-
 export const connect = (app: Express, options: ConnectOptions = {}): void => {
   const sanitize: SanitizeOptions = {
     redactHeaders: options.redactHeaders ? new Set([...options.redactHeaders].map((h) => h.toLowerCase())) : DEFAULT_SANITIZE.redactHeaders,
     redactKeys: options.redactKeys ? new Set([...options.redactKeys].map(normalizeKey)) : DEFAULT_SANITIZE.redactKeys,
     maxBodyBytes: options.maxBodyBytes ?? DEFAULT_SANITIZE.maxBodyBytes,
   };
-  app.use(makeCaptureMiddleware(collector, sanitize));
+  app.use(createCaptureMiddleware(collector, sanitize));
 };
 
 /** Drain this process's captured Examples to `.owl/*.json`. Render later with `the-owl build`. */
