+++ tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: { index: "src/the-owl.ts", vitest: "src/vitest.ts", "bin/cli": "src/bin/cli.ts" },
+  entry: { index: "src/the-owl.ts", vitest: "src/capture/vitest.ts", "bin/cli": "src/bin/cli.ts" },
   format: ["cjs", "esm"],
   dts: true,
   clean: true,
