+++ vitest.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
-  test: { include: ["test/**/*.test.ts"], environment: "node" },
+  test: { include: ["src/**/*.test.ts", "test/**/*.test.ts"], environment: "node" },
 });
